Changes package name for sample plugin
