+++ sample-resource-plugin/src/main/java/org/opensearch/sample/actions/create/CreateSampleResourceRequest.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.security.sample.actions.list;
+package org.opensearch.sample.actions.create;
 
 import java.io.IOException;
 
@@ -14,26 +14,42 @@ import org.opensearch.action.ActionRequest;
 import org.opensearch.action.ActionRequestValidationException;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
+import org.opensearch.sample.Resource;
 
 /**
- * Request object for ListSampleResource transport action
+ * Request object for CreateSampleResource transport action
  */
-public class ListSampleResourceRequest extends ActionRequest {
+public class CreateSampleResourceRequest extends ActionRequest {
 
-    public ListSampleResourceRequest() {}
+    private final Resource resource;
+
+    /**
+     * Default constructor
+     */
+    public CreateSampleResourceRequest(Resource resource) {
+        this.resource = resource;
+    }
 
     /**
      * Constructor with stream input
      * @param in the stream input
      * @throws IOException IOException
      */
-    public ListSampleResourceRequest(final StreamInput in) throws IOException {}
+    public CreateSampleResourceRequest(final StreamInput in) throws IOException {
+        this.resource = new SampleResource(in);
+    }
 
     @Override
-    public void writeTo(final StreamOutput out) throws IOException {}
+    public void writeTo(final StreamOutput out) throws IOException {
+        resource.writeTo(out);
+    }
 
     @Override
     public ActionRequestValidationException validate() {
         return null;
+    }
+
+    public Resource getResource() {
+        return this.resource;
     }
 }
